+++ processor/src/test/resources/expected/Arez_DefaultMethodsModel.java
@@ -69,7 +69,7 @@ public final class Arez_DefaultMethodsModel extends DefaultMethodsModel implemen
         $$arez$$_startedAt = System.currentTimeMillis();
         this.$$arez$$_context.getSpy().reportSpyEvent( new ActionStartedEvent( "DefaultMethodsModel.doStuff", false, new Object[]{time} ) );
       }
-      this.$$arez$$_context.safeProcedure(this.$$arez$$_context.areNamesEnabled() ? "DefaultMethodsModel.doStuff" : null, true, () -> super.doStuff(time) );
+      this.$$arez$$_context.safeAction(this.$$arez$$_context.areNamesEnabled() ? "DefaultMethodsModel.doStuff" : null, true, () -> super.doStuff(time) );
       $$arez$$_completed = true;
       if ( this.$$arez$$_context.areSpiesEnabled() && this.$$arez$$_context.getSpy().willPropagateSpyEvents() ) {
         final long $$arez$$_duration = System.currentTimeMillis() - $$arez$$_startedAt;