+++ processor/src/test/resources/expected/com/example/action/Arez_UnsafeProcedureActionModel.java
@@ -56,7 +56,7 @@ public final class Arez_UnsafeProcedureActionModel extends UnsafeProcedureAction
         $$arez$$_startedAt = System.currentTimeMillis();
         this.$$arez$$_context.getSpy().reportSpyEvent( new ActionStartedEvent( $$arez$$_name() + ".doStuff", false, new Object[]{time} ) );
       }
-      this.$$arez$$_context.procedure(this.$$arez$$_context.areNamesEnabled() ? $$arez$$_name() + ".doStuff" : null, true, () -> super.doStuff(time) );
+      this.$$arez$$_context.action(this.$$arez$$_context.areNamesEnabled() ? $$arez$$_name() + ".doStuff" : null, true, () -> super.doStuff(time) );
       $$arez$$_completed = true;
       if ( this.$$arez$$_context.areSpiesEnabled() && this.$$arez$$_context.getSpy().willPropagateSpyEvents() ) {
         final long $$arez$$_duration = System.currentTimeMillis() - $$arez$$_startedAt;