+++ processor/src/test/resources/expected/com/example/component_name/Arez_ComponentTypeNameModel.java
@@ -61,7 +61,7 @@ public final class Arez_ComponentTypeNameModel extends ComponentTypeNameModel im
         $$arez$$_startedAt = System.currentTimeMillis();
         this.$$arez$$_context.getSpy().reportSpyEvent( new ActionStartedEvent( getComponentName() + ".doStuff", false, new Object[]{time,someOtherParameter} ) );
       }
-      this.$$arez$$_context.safeProcedure(this.$$arez$$_context.areNamesEnabled() ? getComponentName() + ".doStuff" : null, true, () -> super.doStuff(time,someOtherParameter) );
+      this.$$arez$$_context.safeAction(this.$$arez$$_context.areNamesEnabled() ? getComponentName() + ".doStuff" : null, true, () -> super.doStuff(time,someOtherParameter) );
       $$arez$$_completed = true;
       if ( this.$$arez$$_context.areSpiesEnabled() && this.$$arez$$_context.getSpy().willPropagateSpyEvents() ) {
         final long $$arez$$_duration = System.currentTimeMillis() - $$arez$$_startedAt;